Replace legacy VFS with V2
