+++ sqlite3/lib/src/wasm/file_system.dart
@@ -4,12 +4,10 @@ import 'dart:indexed_db';
 import 'dart:math';
 import 'dart:typed_data';
 
-import 'package:path/path.dart' as p;
+import 'package:path/path.dart' as p show url, posix;
 
 import '../../wasm.dart';
 import 'js_interop.dart';
-
-part 'file_system_v2.dart';
 
 /// A virtual file system implementation for web-based `sqlite3` databases.
 abstract class FileSystem {
@@ -38,11 +36,7 @@ abstract class FileSystem {
   void deleteFile(String path);
 
   /// Lists all files stored in this file system.
-  @Deprecated('Use files() instead')
-  List<String> listFiles();
-
-  /// Lists all files stored in this file system.
-  List<String> files();
+  List<String> get files;
 
   /// Deletes all file
   FutureOr<void> clear();
@@ -83,14 +77,6 @@ class FileSystemException implements Exception {
   }
 }
 
-/// An exception thrown by a [FileSystem] implementation when try to access file
-/// outside of persistence root
-class FileSystemAccessException extends FileSystemException {
-  FileSystemAccessException()
-      : super(SqlExtendedError.SQLITE_IOERR_ACCESS,
-            'Path is not within persistence root');
-}
-
 class _InMemoryFileSystem implements FileSystem {
   final Map<String, Uint8List?> _files = {};
 
@@ -98,11 +84,7 @@ class _InMemoryFileSystem implements FileSystem {
   bool exists(String path) => _files.containsKey(path);
 
   @override
-  @Deprecated('Use files() instead')
-  List<String> listFiles() => files();
-
-  @override
-  List<String> files() => _files.keys.toList(growable: false);
+  List<String> get files => _files.keys.toList(growable: false);
 
   @override
   void clear() => _files.clear();
@@ -198,20 +180,23 @@ class _InMemoryFileSystem implements FileSystem {
 ///
 /// As sqlite3's file system is synchronous and IndexedDB isn't, no guarantees
 /// on durability can be made. Instead, file changes are written at some point
-/// after the database is changed.
+/// after the database is changed. However you can wait for changes manually
+/// with [flush]
 ///
 /// In the future, we may want to store individual blocks instead.
 
 class IndexedDbFileSystem implements FileSystem {
-  static const _dbName = 'sqlite3_databases';
-  static const _files = 'files';
-
   final String _persistenceRoot;
+  final String _objectName;
   final Database _database;
-
+  final String dbName;
   final _InMemoryFileSystem _memory = _InMemoryFileSystem();
 
-  IndexedDbFileSystem._(this._persistenceRoot, this._database);
+  bool _closed = false;
+  Future<void>? _current;
+
+  IndexedDbFileSystem._(
+      this._persistenceRoot, this._database, this.dbName, this._objectName);
 
   /// Loads an IndexedDB file system that will consider files in
   /// [persistenceRoot].
@@ -221,64 +206,161 @@ class IndexedDbFileSystem implements FileSystem {
   /// that one [IndexedDbFileSystem] will only see one of them decreases memory
   /// usage.
   ///
-  /// The persistence root can be set to `/` to make all files available.
-  /// Be careful not to use the same or nested [persistenceRoot] for
-  /// different instances. These can overwrite each other and undefined behavior
-  /// can occur.
+  /// [persistenceRoot] is prepended to file names at database level, so you
+  /// have to use relative path in function parameter. Provided paths are
+  /// normalized and resolved like in any operating system:
+  /// /foo/bar/../foo -> /foo/foo
+  /// //foo///bar/../foo/ -> /foo/foo
   ///
-  /// IndexedDbFileSystem doesn't prepend [persistenceRoot] to filenames.
-  /// Rather works more like a guard. If you create/delete file you must prefix
-  /// the path with [persistenceRoot], otherwise saving to IndexedDD will fail
-  /// silently.
-  @Deprecated('Use IndexedDbFileSystemV2 instead')
-  static Future<IndexedDbFileSystem> load(String persistenceRoot) async {
-    // Not using window.indexedDB because we want to support workers too.
-    final database = await self.indexedDB!.open(
-      _dbName,
-      version: 1,
-      onUpgradeNeeded: (event) {
-        final database = event.target.result as Database;
-        database.createObjectStore(_files);
-      },
-    );
-    final fs = IndexedDbFileSystem._(persistenceRoot, database);
+  /// The persistence root can be set to `/` to make all files available.
+  /// Be careful not to use the same or nested [persistenceRoot] for different
+  /// instances with the same database and object name. These can overwrite each
+  /// other and undefined behavior can occur.
+  ///
+  /// With [dbName] you can set IndexedDB database name
+  /// With [objectName] you can set file storage object key name
+  static Future<IndexedDbFileSystem> init({
+    String persistenceRoot = '/',
+    String dbName = 'sqlite3_databases',
+    String objectName = 'files',
+  }) async {
+    final openDatabase = (int? version) {
+      // Not using window.indexedDB because we want to support workers too.
+      return self.indexedDB!.open(
+        dbName,
+        version: version,
+        onUpgradeNeeded: version == null
+            ? null
+            : (event) {
+                final database = event.target.result as Database;
+                database.createObjectStore(objectName);
+              },
+      );
+    };
 
-    // Load persisted files from IndexedDb
-    final transaction = database.transactionStore(_files, 'readonly');
-    final files = transaction.objectStore(_files);
-
-    await for (final entry in files.openCursor(autoAdvance: true)) {
-      final path = entry.primaryKey! as String;
-
-      if (p.url.isWithin(persistenceRoot, path)) {
-        final object = await entry.value as Blob?;
-        if (object == null) continue;
-
-        fs._memory._files[path] = await object.arrayBuffer();
-      }
+    // Check if a new objectName is used on existing database. Must run
+    // upgrade in this case.
+    // Because of a bug in DartVM, it can run into a deadlock when run parallel
+    // access to the same database while upgrading
+    // https://github.com/dart-lang/sdk/issues/48854
+    var database = await openDatabase(null);
+    if (!(database.objectStoreNames ?? []).contains(objectName)) {
+      database.close();
+      database = await openDatabase((database.version ?? 1) + 1);
     }
 
+    final root = p.posix.normalize('/$persistenceRoot');
+    final fs = IndexedDbFileSystem._(root, database, dbName, objectName);
+    await fs._sync();
     return fs;
   }
 
-  bool _shouldPersist(String path) =>
-      path.startsWith('/tmp/') || p.url.isWithin(_persistenceRoot, path);
+  /// Returns all database
+  /// Returns null if 'IndexedDB.databases()' function is not supported in the
+  /// JS engine
+  static Future<List<DatabaseName>?> databases() {
+    return self.indexedDB!.databases();
+  }
 
-  void _canPersist(String path) {
-    if (!_shouldPersist(path)) {
-      throw FileSystemAccessException();
+  static Future<void> deleteDatabase(
+      [String dbName = 'sqlite3_databases']) async {
+    // The deadlock issue can cause problem here too
+    // https://github.com/dart-lang/sdk/issues/48854
+    await self.indexedDB!.deleteDatabase(dbName);
+  }
+
+  bool get isClosed => _closed;
+
+  Future<void> close() async {
+    if (!_closed) {
+      await flush();
+      _database.close();
+      _memory.clear();
+      _closed = true;
+    }
+  }
+
+  void _checkClosed() {
+    if (_closed) {
+      throw FileSystemException(SqlError.SQLITE_IOERR, 'FileSystem closed');
+    }
+  }
+
+  /// Flush and reload files from IndexedDB
+  Future<void> sync() async {
+    _checkClosed();
+    await _mutex(() => _sync());
+  }
+
+  Future<void> _sync() async {
+    final transaction = _database.transactionStore(_objectName, 'readonly');
+    final files = transaction.objectStore(_objectName);
+    _memory.clear();
+    await for (final entry in files.openCursor(autoAdvance: true)) {
+      final path = entry.primaryKey! as String;
+      if (p.url.isWithin(_persistenceRoot, path)) {
+        final object = await entry.value as Blob?;
+        if (object == null) continue;
+
+        _memory._files[_relativePath(path)] = await object.arrayBuffer();
+      }
+    }
+  }
+
+  String _normalize(String path) {
+    if (path.endsWith('/') || path.endsWith('.')) {
+      throw FileSystemException(
+          SqlExtendedError.SQLITE_CANTOPEN_ISDIR, 'Path is a directory');
+    }
+    return p.posix.normalize('/${path}');
+  }
+
+  /// Returns the absolute path to IndexedDB
+  String absolutePath(String path) {
+    _checkClosed();
+    return p.posix.normalize('/$_persistenceRoot/$path');
+  }
+
+  String _relativePath(String path) =>
+      p.posix.normalize(path.replaceFirst(_persistenceRoot, '/'));
+
+  Future<void> _mutex(Future<void> Function() body) async {
+    await flush();
+    try {
+      _current = body();
+      await _current;
+    } on Exception catch (e, s) {
+      print(e);
+      print(s);
+    } finally {
+      _current = null;
+    }
+  }
+
+  /// Waits for pending IndexedDB operations
+  Future<void> flush() async {
+    _checkClosed();
+    if (_current != null) {
+      try {
+        await Future.wait([_current!]);
+      } on Exception catch (_) {
+      } finally {
+        _current = null;
+      }
     }
   }
 
   void _writeFileAsync(String path) {
-    if (_shouldPersist(path)) {
-      Future.sync(() async {
-        final transaction = _database.transaction(_files, 'readwrite');
-        await transaction
-            .objectStore(_files)
-            .put(Blob(<Uint8List>[_memory._files[path] ?? Uint8List(0)]), path);
-      });
-    }
+    Future.sync(() async {
+      await _mutex(() => _writeFile(path));
+    });
+  }
+
+  Future<void> _writeFile(String path) async {
+    final transaction = _database.transaction(_objectName, 'readwrite');
+    await transaction.objectStore(_objectName).put(
+        Blob(<Uint8List>[_memory._files[path] ?? Uint8List(0)]),
+        absolutePath(path));
   }
 
   @override
@@ -287,89 +369,98 @@ class IndexedDbFileSystem implements FileSystem {
     bool errorIfNotExists = false,
     bool errorIfAlreadyExists = false,
   }) {
-    _canPersist(path);
-    final exists = _memory.exists(path);
-    _memory.createFile(path, errorIfAlreadyExists: errorIfAlreadyExists);
+    _checkClosed();
+    final _path = _normalize(path);
+    final exists = _memory.exists(_path);
+
+    _memory.createFile(
+      _path,
+      errorIfAlreadyExists: errorIfAlreadyExists,
+      errorIfNotExists: errorIfNotExists,
+    );
 
     if (!exists) {
       // Just created, so write
-      _writeFileAsync(path);
+      _writeFileAsync(_path);
     }
   }
 
   @override
   String createTemporaryFile() {
-    var i = 0;
-    while (exists('/tmp/$i')) {
-      i++;
-    }
-
-    final fileName = '/tmp/$i';
-    createFile(fileName);
-    return fileName;
+    _checkClosed();
+    final path = _memory.createTemporaryFile();
+    _writeFileAsync(path);
+    return path;
   }
 
   @override
   void deleteFile(String path) {
-    _canPersist(path);
-    _memory.deleteFile(path);
-
-    if (_shouldPersist(path)) {
-      Future.sync(() async {
-        final transaction = _database.transactionStore(_files, 'readwrite');
-        await transaction.objectStore(_files).delete(path);
-      });
-    }
+    _checkClosed();
+    final _path = _normalize(path);
+    _memory.deleteFile(_path);
+    Future.sync(
+      () => _mutex(() async {
+        final transaction =
+            _database.transactionStore(_objectName, 'readwrite');
+        await transaction.objectStore(_objectName).delete(absolutePath(_path));
+      }),
+    );
   }
 
-  /// Deletes all file
   @override
-  void clear() {
-    final dbFiles = files();
+  Future<void> clear() async {
+    _checkClosed();
+    final _files = files;
     _memory.clear();
-    Future.sync(() async {
-      final transaction = _database.transactionStore(_files, 'readwrite');
-      for (final file in dbFiles) {
-        await transaction.objectStore(_files).delete(file);
+    await _mutex(() async {
+      final transaction = _database.transactionStore(_objectName, 'readwrite');
+      for (final file in _files) {
+        final f = absolutePath(file);
+        await transaction.objectStore(_objectName).delete(f);
       }
     });
   }
 
   @override
   bool exists(String path) {
-    _canPersist(path);
-    return _memory.exists(path);
+    _checkClosed();
+    final _path = _normalize(path);
+    return _memory.exists(_path);
   }
 
   @override
-  List<String> listFiles() => files();
-
-  @override
-  List<String> files() => _memory.files();
+  List<String> get files {
+    _checkClosed();
+    return _memory.files;
+  }
 
   @override
   int read(String path, Uint8List target, int offset) {
-    _canPersist(path);
-    return _memory.read(path, target, offset);
+    _checkClosed();
+    final _path = _normalize(path);
+    return _memory.read(_path, target, offset);
   }
 
   @override
   int sizeOfFile(String path) {
-    _canPersist(path);
-    return _memory.sizeOfFile(path);
+    _checkClosed();
+    final _path = _normalize(path);
+    return _memory.sizeOfFile(_path);
   }
 
   @override
   void truncateFile(String path, int length) {
-    _canPersist(path);
-    _memory.truncateFile(path, length);
-    _writeFileAsync(path);
+    _checkClosed();
+    final _path = _normalize(path);
+    _memory.truncateFile(_path, length);
+    _writeFileAsync(_path);
   }
 
   @override
   void write(String path, Uint8List bytes, int offset) {
-    _canPersist(path);
-    _memory.write(path, bytes, offset);
-    _writeFileAsync(path);
+    _checkClosed();
+    final _path = _normalize(path);
+    _memory.write(_path, bytes, offset);
+    _writeFileAsync(_path);
   }
 }
