+++ sqlite3/test/wasm/file_system_test.dart
@@ -15,30 +15,13 @@ Future<void> main() async {
     _testWith(FileSystem.inMemory);
   });
 
-  group('indexed db with legacy vfs', () {
-    _testWith(() => IndexedDbFileSystem.load(_fsRoot));
-    _testAccess(() => IndexedDbFileSystem.load(_fsRoot));
-  });
-
-  group('indexed db with vfs v2.0', () {
-    _testWith(() => IndexedDbFileSystemV2.init(
+  group('indexed db', () {
+    _testWith(() => IndexedDbFileSystem.init(
         persistenceRoot: _fsRoot, dbName: _randomName()));
-    _testV2Persistence();
+    _testPersistence();
   });
 
-  group('basic persistence', () {
-    test('basic persistence V1', () async {
-      final fs = await IndexedDbFileSystem.load(_fsRoot);
-      await _basicPersistence(fs);
-    });
-
-    test('basic persistence V2', () async {
-      final fs = await IndexedDbFileSystemV2.init(dbName: _randomName());
-      await _basicPersistence(fs);
-    });
-  });
-
-  group('vfs v2.0 path normalization', () {
+  group('path normalization', () {
     _testPathNormalization();
   });
 }
@@ -47,29 +30,21 @@ final _random = Random(DateTime.now().millisecond);
 String _randomName() => _random.nextInt(0x7fffffff).toString();
 
 Future<void> _disposeFileSystem(FileSystem fs) async {
-  if (fs is IndexedDbFileSystemV2) {
+  if (fs is IndexedDbFileSystem) {
     await fs.close();
-    await IndexedDbFileSystemV2.deleteDatabase(fs.dbName);
+    await IndexedDbFileSystem.deleteDatabase(fs.dbName);
   } else {
     fs.clear();
   }
 }
 
 Future<void> _runPathTest(String root, String path, String resolved) async {
-  final fs = await IndexedDbFileSystemV2.init(
+  final fs = await IndexedDbFileSystem.init(
       persistenceRoot: root, dbName: _randomName());
   fs.createFile(path);
-  final absPath = fs.absolutePath(fs.files().first);
+  final absPath = fs.absolutePath(fs.files.first);
   expect(absPath, resolved);
   await _disposeFileSystem(fs);
-}
-
-Future<void> _basicPersistence(FileSystem fs) async {
-  fs.createFile('$_fsRoot/test');
-  expect(fs.exists('$_fsRoot/test'), true);
-  await Future<void>.delayed(const Duration(milliseconds: 1000));
-  final fs2 = await IndexedDbFileSystem.load(_fsRoot);
-  expect(fs2.exists('$_fsRoot/test'), true);
 }
 
 Future<void> _testPathNormalization() async {
@@ -82,7 +57,7 @@ Future<void> _testPathNormalization() async {
     await _runPathTest('../', 'test', '/test');
   });
 
-  test('normalization', () async {
+  test('path', () async {
     await _runPathTest('', 'test', '/test');
     await _runPathTest('/', '../test', '/test');
     await _runPathTest('/', '../test/../../test', '/test');
@@ -115,32 +90,32 @@ Future<void> _testWith(FutureOr<FileSystem> Function() open) async {
 
   test('can create files', () {
     expect(fs.exists('$_fsRoot/foo.txt'), isFalse);
-    expect(fs.files(), isEmpty);
+    expect(fs.files, isEmpty);
     fs.createFile('$_fsRoot/foo.txt');
     expect(fs.exists('$_fsRoot/foo.txt'), isTrue);
-    expect(fs.files(), ['$_fsRoot/foo.txt']);
+    expect(fs.files, ['$_fsRoot/foo.txt']);
     fs.deleteFile('$_fsRoot/foo.txt');
-    expect(fs.files(), isEmpty);
+    expect(fs.files, isEmpty);
   });
 
   test('can create and delete multiple files', () {
     for (var i = 1; i <= 10; i++) {
       fs.createFile('$_fsRoot/foo$i.txt');
     }
-    expect(fs.files(), hasLength(10));
-    for (final f in fs.files()) {
+    expect(fs.files, hasLength(10));
+    for (final f in fs.files) {
       fs.deleteFile(f);
     }
-    expect(fs.files(), isEmpty);
+    expect(fs.files, isEmpty);
   });
 
   test('can create files and clear fs', () {
     for (var i = 1; i <= 10; i++) {
       fs.createFile('$_fsRoot/foo$i.txt');
     }
-    expect(fs.files(), hasLength(10));
+    expect(fs.files, hasLength(10));
     fs.clear();
-    expect(fs.files(), isEmpty);
+    expect(fs.files, isEmpty);
   });
 
   test('reads and writes', () {
@@ -163,39 +138,14 @@ Future<void> _testWith(FutureOr<FileSystem> Function() open) async {
   });
 }
 
-Future<void> _testAccess(Future<IndexedDbFileSystem> Function() open) async {
-  late IndexedDbFileSystem fs;
-
-  setUp(() async => fs = await open());
-  tearDown(() => _disposeFileSystem(fs));
-
-  test('access permissions', () {
-    expect(() => fs.exists('/test2/foo.txt'),
-        throwsA(isA<FileSystemAccessException>()));
-    expect(() => fs.createFile('/test2/foo.txt'),
-        throwsA(isA<FileSystemAccessException>()));
-    expect(() => fs.write('/test2/foo.txt', Uint8List(0), 0),
-        throwsA(isA<FileSystemAccessException>()));
-    expect(() => fs.sizeOfFile('/test2/foo.txt'),
-        throwsA(isA<FileSystemAccessException>()));
-    expect(() => fs.read('/test2/foo.txt', Uint8List(0), 0),
-        throwsA(isA<FileSystemAccessException>()));
-    expect(() => fs.truncateFile('/test2/foo.txt', 0),
-        throwsA(isA<FileSystemAccessException>()));
-    expect(() => fs.deleteFile('/test2/foo.txt'),
-        throwsA(isA<FileSystemAccessException>()));
-    expect(fs.createTemporaryFile(), '/tmp/0');
-  });
-}
-
-void _testV2Persistence() {
-  test('advanced persistence', () async {
+void _testPersistence() {
+  test('Properly persist into IndexedDB', () async {
     final data = Uint8List.fromList([for (var i = 0; i < 255; i++) i]);
     final dbName = _randomName();
 
     await expectLater(
       () async {
-        final databases = (await IndexedDbFileSystemV2.databases())
+        final databases = (await IndexedDbFileSystem.databases())
             ?.map((e) => e.name)
             .toList();
         return databases == null ? true : !databases.contains(dbName);
@@ -204,17 +154,17 @@ void _testV2Persistence() {
       reason: 'There must be no database named dbName at the beginning',
     );
 
-    final db1 = await IndexedDbFileSystemV2.init(dbName: dbName);
-    expect(db1.files().length, 0, reason: 'db1 is not empty');
+    final db1 = await IndexedDbFileSystem.init(dbName: dbName);
+    expect(db1.files.length, 0, reason: 'db1 is not empty');
 
     db1.createFile('test');
     db1.write('test', data, 0);
     await db1.flush();
-    expect(db1.files().length, 1, reason: 'There must be only one file in db1');
+    expect(db1.files.length, 1, reason: 'There must be only one file in db1');
     expect(db1.exists('test'), true, reason: 'The test file must exist in db1');
 
-    final db2 = await IndexedDbFileSystemV2.init(dbName: dbName);
-    expect(db2.files().length, 1, reason: 'There must be only one file in db2');
+    final db2 = await IndexedDbFileSystem.init(dbName: dbName);
+    expect(db2.files.length, 1, reason: 'There must be only one file in db2');
     expect(db2.exists('test'), true, reason: 'The test file must exist in db2');
 
     final read = Uint8List(255);
@@ -223,10 +173,10 @@ void _testV2Persistence() {
         reason: 'The data written and read do not match');
 
     await db2.clear();
-    expect(db2.files().length, 0, reason: 'There must be no files in db2');
-    expect(db1.files().length, 1, reason: 'There must be only one file in db1');
+    expect(db2.files.length, 0, reason: 'There must be no files in db2');
+    expect(db1.files.length, 1, reason: 'There must be only one file in db1');
     await db1.sync();
-    expect(db1.files().length, 0, reason: 'There must be no files in db1');
+    expect(db1.files.length, 0, reason: 'There must be no files in db1');
 
     await db1.close();
     await db2.close();
@@ -234,11 +184,11 @@ void _testV2Persistence() {
     await expectLater(() => db1.sync(), throwsA(isA<FileSystemException>()));
     await expectLater(() => db2.sync(), throwsA(isA<FileSystemException>()));
 
-    await IndexedDbFileSystemV2.deleteDatabase(dbName);
+    await IndexedDbFileSystem.deleteDatabase(dbName);
 
     await expectLater(
       () async {
-        final databases = (await IndexedDbFileSystemV2.databases())
+        final databases = (await IndexedDbFileSystem.databases())
             ?.map((e) => e.name)
             .toList();
         return databases == null ? true : !databases.contains(dbName);
